switch to module type
