+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "homepage": "https://jz6.github.io/RatingEpicGames",
   "private": true,
+  "type": "module",
   "dependencies": {
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
